Reorganizing everything into less
A lot of changes here. Moving Bootstrap to the vendor folder and using their less instead of compiled CSS. Added a custom variables file for easier control over Bootstrap, and created a master production.less that can be compiled to /bithub/production.css - Should be good to go for more maintainable Less!

Signed-off-by: Tom Greever <anybody@tomgreever.com>